Rearrange pattern rules in Makefile to fix Mac OS
The sets of patterns to match certain source files for different CPU
features are based on suffixes in their filenames. make on Linux seems
to pick the most specific rule whereas make shipped with Xcode does
pick the first and therefore disregards the remaining rules. This was
fixed by rearranging the rules from most specific to most general.

See issue #4