Fix/improve creation and removing of bin folder
The "bin" folder was never created explicitly in the Makefile therefore
this was added in the according target. This was propably never noticed
as the "clean" target did not remove the "bin" folder once created (it
only removed its contents). To make sure this gets noticed and to get a
"cleaner" state after executing "make clean" the "bin" folder now gets
removed completely.