Add support for macOS
This is a rework of PR #4 which was split into smaller PRs which where
not related to macOS only. This PR waschanged to reflect the changes which
were made in the mean time.

Compiling on macOS requires following changes to the Makefile:

Some checks were introduced to set compiler flags and file extensions
required or specific for macOS.

Additionally the order of general pattern rules was changed.
The sets of patterns to match certain source files for different CPU
features are based on suffixes in their filenames. make on Linux seems
to pick the most specific rule whereas make shipped with Xcode does
pick the first and therefore disregards the remaining rules. This was
fixed by rearranging the rules from most specific to most general.
This potentially fixes issue #10.