Allow for defining SHA1DC_BIGENDIAN externally
In 15c90b2 the SHA1DC_BIGENDIAN macro would be undefined if it was
already defined. It's not explained why, and I can't think of a reason
for why.

This allows for doing -DSHA1DC_BIGENDIAN on a platform that's not
listed here. It would be nice e.g. for the git project if someone
running on such a platform didn't have to monkeypatch the code,
instead they can just provide a -DSHA1DC_BIGENDIAN to the Makefile.

I removed the closing ifdef comment because I thought now that it's
nested this just became less clear, and since this whole thing just
spans 9 lines it's more clearly delineated by whitespace.