.gitignore: ignore lib/.depend/ and other .depend/ directories
This is a hack so that I can run "make" cleanly in git.git. This is
really a wart around there being no easy way to have tracked
.gitignore changes to submodules.

You can hack e.g. .git/modules/sha1collisiondetection/info/exclude,
but that's not under source control, and thus needs to be done by the
Makefile.

This is an easy workaround that won't show the module dirty everytime
"make" is run, while we think about some longer-term fix for this.