Big Endian Detection: fix breakage on *BSD & newlib
In 33a694a ("Fix issues with a big endian platform", 2017-05-15) *BSD
and newlib was broken because _BIG_ENDIAN is always defined, even on
little-endian platforms.

To fix this we need to also check the single-underscore version of
BIG_ENDIAN & LITTLE_ENDIAN, but only do this if _LITTLE_ENDIAN is also
defined. This makes sure that on Solaris we don't break where
_BIG_ENDIAN is simply defined, not to any value (but _LITTLE_ENDIAN
won't be). See
http://src.illumos.org/source/xref/illumos-gate/usr/src/uts/common/sys/isa_defs.h#397

The https://github.com/cr-marcstevens/sha1collisiondetection/pull/34
pull request has a lot more details on the various issues involved,
things that were tried (and not tried) etc.

[Commit message & rebase by Ævar]