sha1.c: replace white/blacklist with allow/denylist
The words "whitelist" and "blacklist" have cultural implications that
are not inclusive. The words "allowlist" and "denylist" are
self-descriptive without requiring cultural implications to understand
the words in context.

A similar set of changes were sent to the Git project, but in those
cases the words could be removed by rephrasing the sentences that used
them, increasing clarity. In this example, the terms are used concretely
enough that rephrasing would decrease clarity. Thus, a relatively
mechanical replacement was applied.

These changes only affect comment lines, nothing functional.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>